[REFACTOR] Place Entity 수정 및 Place DVO 생성
